Changed Fields To Float

+++ Headers/Scene/material.h
@@ -9,11 +9,11 @@ struct Material
         glm::vec4 emissionColor;
         glm::vec4 baseColor;
         
-        int reflectance;
-        int roughness;
-        int metallic;
+        float reflectance;
+        float roughness;
+        float metallic;
         
-        Material(glm::vec3 baseColor = glm::vec3(0), glm::vec3 emissionColor = glm::vec3(0), int metallic = 0, int roughness = 0, int reflectance = 0)
+        Material(glm::vec3 baseColor = glm::vec3(0), glm::vec3 emissionColor = glm::vec3(0), float metallic = 0, float roughness = 0, float reflectance = 0)
         {
             this->emissionColor = glm::vec4(emissionColor, 1.0f);
             this->baseColor = glm::vec4(baseColor, 1.0f);
@@ -24,7 +24,7 @@ struct Material
 
     private:
         // This is only here to make the material size divisable by 4 to work with std430 rules
-        int padding = 0;
+        float padding = 0;
 };
 
 #endif
